requirements.txt
PyOTA
flask
flask_cors
pytest
python-dotenv